// src/constants/api/enums/module_names.ts
export enum MODULE {
  API_LOOKUP_ENTRIES = "api.lookup_entries",
  API_LOOKUP_VERSIONS = "api.lookup_versions",
  API_ORDER_HISTORIES = "api.order_histories",
  API_ORDERS_DATA = "api.orders_data",
  API_PRODUCT_TYPES = "api.product_types",
  API_ROLE_MAPPINGS = "api.role_mappings",
  API_USERS = "api.users",
  API_EDIT_PAGE_PAX_TAG = "api.edit_page_pax_tag",
  API_CLONE_PAGE_PAX_TAG = "api.clone_page_pax_tag",
}